requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
openai==1.59.6
numpy==1.26.4
pandas==2.2.3
tiktoken==0.8.0
faiss-cpu==1.9.0.post1
streamlit==1.41.1
requests==2.32.3
rich==13.9.4
